+++ src/star_discovery/recovery/nodes/abc/html_base.py
@@ -26,16 +26,20 @@ class HTMLBaseNode(BaseNode, ABC):
 
     def __init__(
         self,
+        depth: int,
         parent: HTMLElementBaseNode | None,
         elm: NavigableString | Tag,
         index: int = 0,
     ):
         self._elm = elm
         self._index = index
-        super().__init__(parent)
+        super().__init__(depth, parent)
 
     @override
     def _path_segment_value(self) -> str:
         if self._index:
             return f"{self._index}:{self._value}"
         return self._value
+
+    def max_depth(self) -> int:
+        raise NotImplementedError("max_depth", self)